Refactor course and chapter management to use VO classes; update service and controller methods for consistency; enhance section handling with new resource types and UI components

+++ src/main/java/com/zhangziqi/online_course_mine/service/impl/CategoryServiceImpl.java
@@ -318,11 +318,11 @@ public class CategoryServiceImpl implements CategoryService {
         
         // 获取课程数量
         long courseCount = categoryRepository.countCoursesByCategoryId(category.getId());
-        vo.setCourseCount(courseCount);
+        vo.setCourseCount((int) courseCount);
         
         // 获取子分类数量
         long childrenCount = categoryRepository.countByParentId(category.getId());
-        vo.setChildrenCount(childrenCount);
+        vo.setChildrenCount((int) childrenCount);
         
         return vo;
     }
